refactor(page): split HeatMap components to ColorLegend, DrawingTools, HeatmapLayer, MapCanvas, MapNavigator, StatCards, and Toolbar
- add ColorLegend for displaying color gradients based on data range
- add DrawingTools for bounding box and polygon drawing functionalities
- add HeatmapLayer to render heatmap based on provided bindings and grid size
- add MapCanvas to encapsulate map rendering and loading states
- add MapNavigator and MapFitter for map navigation and fitting to bounding boxes
- add StatCards to display statistical data with loading skeletons
- add Toolbar for user interaction with drawing modes and clearing selections.
